+++ Blurred-ios/Cells/SearchCell.swift
@@ -1,0 +1,14 @@
+//
+//  SearchCell.swift
+//  Blurred-ios
+//
+//  Created by Martin Velev on 6/28/20.
+//  Copyright © 2020 BlurrMC. All rights reserved.
+//
+
+import UIKit
+
+class SearchCell: UICollectionViewCell {
+    // Put the prepare for reuse here.
+}
+