+++ Blurred-ios/SearchViewController.swift
@@ -8,15 +8,17 @@
 
 import UIKit
 
-class SearchViewController: UIViewController, UISearchBarDelegate {
+class SearchViewController: UIViewController { // , UICollectionViewDataSource, UISearchBarDelegate
+    
+    
     
     
     // Communicates with the api for search results
     override func viewDidLoad() {
         super.viewDidLoad()
-        self.searchBarTextField?.delegate? = self
         // Do any additional setup after loading the view.
     }
+    @IBOutlet weak var collectionView: UICollectionView!
     @IBOutlet weak var searchBarTextField: UISearchBar!