+++ CHANGELOG.md
@@ -20,6 +20,9 @@
 * Fixed bug where done button in upload details would take you to the details of the previous video you uploaded.
 * Fixed bug where app would crash if you recorded video, went to playback, and went back to video it would crash the app.
 * Fixed bug with cells not being reused in channel video thumbnails.
+* Added groups for better sorting
+* Added collectionview (for results) to the search.
+* Added back the record icon in tab bar.
 ## v0.3 (PRERELEASE)
 * Changed scale factor of avatar image in channel view.
 * Added description field